+++ banyand/measure/part.go
@@ -45,6 +45,7 @@ const (
 	measureSeriesMetadataName = "smeta"
 
 	metadataFilename               = "metadata.json"
+	tagTypeFilename                = "tag.type"
 	primaryFilename                = measurePrimaryName + ".bin"
 	metaFilename                   = measureMetaName + ".bin"
 	timestampsFilename             = measureTimestampsName + ".bin"
@@ -112,6 +113,7 @@ func openMemPart(mp *memPart) *part {
 type memPart struct {
 	tagFamilyMetadata map[string]*bytes.Buffer
 	tagFamilies       map[string]*bytes.Buffer
+	tagType           tagType
 	meta              bytes.Buffer
 	primary           bytes.Buffer
 	timestamps        bytes.Buffer
@@ -140,6 +142,11 @@ func (mp *memPart) mustCreateMemTagFamilyWriters(name string) (fs.Writer, fs.Wri
 
 func (mp *memPart) reset() {
 	mp.partMetadata.reset()
+	if mp.tagType == nil {
+		mp.tagType = make(tagType)
+	} else {
+		mp.tagType.reset()
+	}
 	mp.meta.Reset()
 	mp.primary.Reset()
 	mp.timestamps.Reset()
@@ -200,7 +207,7 @@ func (mp *memPart) mustInitFromDataPoints(dps *dataPoints) {
 		uncompressedBlockSizeBytes += uncompressedDataPointSizeBytes(i, dps)
 	}
 	bsw.MustWriteDataPoints(sidPrev, dps.timestamps[indexPrev:], dps.versions[indexPrev:], dps.tagFamilies[indexPrev:], dps.fields[indexPrev:])
-	bsw.Flush(&mp.partMetadata)
+	bsw.Flush(&mp.partMetadata, &mp.tagType)
 	releaseBlockWriter(bsw)
 }
 
@@ -226,6 +233,7 @@ func (mp *memPart) mustFlush(fileSystem fs.FileSystem, path string) {
 		fs.MustFlushAtomic(fileSystem, mp.seriesMetadata.Buf, filepath.Join(path, seriesMetadataFilename), storage.FilePerm)
 	}
 
+	mp.tagType.mustWriteTagType(fileSystem, path)
 	mp.partMetadata.mustWriteMetadata(fileSystem, path)
 	// No SyncPath: mustWriteMetadata goes through WriteAtomic which already
 	// fsyncs the parent directory after rename, covering the dirent changes
@@ -413,6 +421,19 @@ func CreatePartFileReaderFromPath(partPath string, lfs fs.FileSystem) ([]queue.F
 	for _, e := range ee {
 		if e.IsDir() {
 			continue
+		}
+
+		if e.Name() == tagTypeFilename {
+			tagTypePath := path.Join(partPath, e.Name())
+			tagTypeReader, openErr := lfs.OpenFile(tagTypePath)
+			if openErr != nil {
+				logger.Panicf("cannot open tag type file %q: %s", tagTypePath, openErr)
+			}
+			readers = append(readers, tagTypeReader)
+			files = append(files, queue.FileInfo{
+				Name:   tagTypeFilename,
+				Reader: tagTypeReader.SequentialRead(),
+			})
 		}
 
 		// Tag family metadata files (.tfm)